requirements.txt
channels==0.17.3
Django==1.10.1
django-nested-admin==3.0.8
django-password-reset==0.9
django-datetime-widget==0.9.3
mysqlclient==1.3.7
pusher==1.3.0
Pillow==3.3.0
service_identity==16.0.0
reportlab==3.3.0
pyexcel-xls==0.3.1
asgi_redis==1.3.0
Django-admin-tools==0.8.0